Moves `ApplicativeMappableN` and `ContainerN` interfaces from `interfaces.aliases` to `interfaces` module
