+++ returns/interfaces/applicative_mappable.py
