+++ returns/interfaces/specific/io.py
@@ -1,7 +1,7 @@
 from abc import abstractmethod
 from typing import TYPE_CHECKING, Callable, NoReturn, Type, TypeVar
 
-from returns.interfaces.aliases import container
+from returns.interfaces import container
 from returns.primitives.hkt import KindN
 
 if TYPE_CHECKING:
