+++ returns/interfaces/specific/result.py
@@ -8,8 +8,7 @@ For impure result see
 from abc import abstractmethod
 from typing import TYPE_CHECKING, Callable, NoReturn, Type, TypeVar
 
-from returns.interfaces import altable, rescuable, unwrappable
-from returns.interfaces.aliases import container
+from returns.interfaces import altable, container, rescuable, unwrappable
 from returns.primitives.hkt import KindN
 
 if TYPE_CHECKING:
