+++ returns/maybe.py
@@ -17,7 +17,7 @@ from typing_extensions import final
 
 from returns._generated.iterable import iterable_kind
 from returns.interfaces import rescuable, unwrappable
-from returns.interfaces.aliases.container import Container1
+from returns.interfaces.container import Container1
 from returns.primitives.container import BaseContainer
 from returns.primitives.exceptions import UnwrapFailedError
 from returns.primitives.hkt import Kind1, SupportsKind1
